Fix cookies not appearing in SFAuthenticationSession by setting them with an expiry time
Session cookies (those set without an expires or max-age param) do not
appear in SFAuthenticationSession sessions, even if the session is
still active in Safari and the cookies are readable there. I don’t know
if this is a bug or intentional.

Setting an expiry time on the cookie via expires or max-age fixes the
issue and the cookies are available in the SFAuthenticationSession session.